src/DORECO_back/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.6.0
PyMySQL==1.1.1
cryptography==43.0.3
Pillow==11.0.0
qrcode==8.0
python-dotenv==1.0.1 